+++ src/lib.rs
@@ -1,5 +1,6 @@
 mod app;
 mod color;
+mod picker;
 
 pub use app::ColorPicker as Epick;